Update delete method and implement notify toastr

+++ app/Http/Controllers/CargoController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 
 use App\Cargo;
+use App\CargoColaborador;
 use Illuminate\Http\Request;
 
 class CargoController extends Controller
@@ -34,8 +35,8 @@ class CargoController extends Controller
         $request['cpf'] = str_replace(array('.','-','/'), "", $request->cpf);
         $this->cargo->fill($request->input());
         $this->cargo->push();
-
-        return redirect()->route('CargoShow')->with('success', 'Cargo criado com sucesso.');
+        toastr()->success('Cargo criado com sucesso','Operação realizada');
+        return redirect()->route('CargoShow');
     }
 
     public function update(Request $request)
@@ -52,13 +53,18 @@ class CargoController extends Controller
     public static function destroy(Request $request)
     {
         $cargo = cargo::find($request->id);
-
-        if ($cargo) {
-            $cargo->delete();
-            return redirect()->route('CargoShow')->with('success', 'Registro excluído com sucesso.');
-        } else {
-            return redirect()->route('CargoShow')->with('error', 'Registro não encontrado.');
+        if (CargoColaborador::checkBondWithCargo($request->id)) {
+            if ($cargo) {
+                $cargo->delete();
+                toastr()->success( 'Registro excluído com sucesso','Operação realizada');
+                return redirect()->route('CargoShow');
+            } else {
+                toastr()->error( 'Registro não encontrado','Falha na operação');
+                return redirect()->route('CargoShow');
+            }
         }
+        toastr()->error('Registro possui vinculo ativo com colaboradores na tabela Cargo Colaborador', 'Falha na operação');
+        return redirect()->route('CargoShow');
     }
